Header, menu and toggle styling compiled to header.styles. GhostContainer moved to icons folder
