+++ src/components/Header.jsx
@@ -3,11 +3,11 @@ import { Link } from "react-router-dom";
 
 // Components
 import { HamburgerMenu, StyledHeader } from "./Header.styles";
-import { GhostContainer, GhostWrapper } from "./GhostContainer.styles";
-import { ToggleDiv, MenuOverlay, StyledMenuLink, NavContainer, ToggleWrapper } from "./MenuOverlay.styles";
+import { ToggleDiv, MenuOverlay, StyledMenuLink, NavContainer, ToggleWrapper } from "./Header.styles";
 import Toggle from "./Toggle";
 
 // Icons
+import { GhostContainer, GhostWrapper } from "./icons/GhostContainer.styles"
 import GhostIcon from "./icons/Ghost";
 
 // Data / Language
