+++ src/pages/ChooseDifficulty.jsx
@@ -4,7 +4,7 @@ import { useNavigate } from "react-router-dom";
 import { HeadingCard, PointCard } from "../components/cards/Cards.styles";
 import { BigButton } from "../components/buttons/Buttons.styles";
 import { BigButtonsCard } from "../components/cards/Cards.styles";
-import { GhostContainer, GhostWrapper } from "../components/GhostContainer.styles";
+import { GhostContainer, GhostWrapper } from "../components/icons/GhostContainer.styles";
 import { ErrorModal } from "../components/ErrorModal";
 
 // Icons
