+++ src/pages/Home.jsx
@@ -4,7 +4,7 @@ import { useState, useEffect } from "react";
 // Components
 import { WideButton } from "../components/buttons/Buttons.styles";
 import { BaseCard } from "../components/cards/Cards.styles";
-import { GhostContainer, GhostWrapper } from "../components/GhostContainer.styles";
+import { GhostContainer, GhostWrapper } from "../components/icons/GhostContainer.styles";
 import { RegisteredPlayersCard } from "../components/cards/Cards.styles";
 import { ErrorModal } from "../components/ErrorModal";
